解析数据转发命令

+++ src/main/java/com/bms/industry/service/BusSiteService.java
@@ -101,7 +101,7 @@ public class BusSiteService {
     }
 
     @Transactional(readOnly = true)
-    public BusSite findByOId(Long oId) {
+    public BusSite findByOId(Integer oId) {
         return busSiteRepository.findByoId(oId);
     }
 }
